move things around, test timeout parsing
