+++ failgood/jvm/src/failgood/internal/util/Getenv.kt
@@ -1,3 +1,3 @@
-package failgood.util
+package failgood.internal.util
 
 internal fun getenv(name: String): String? = System.getenv(name)
