+++ failgood/jvm/src/failgood/internal/util/StringUtils.kt
@@ -1,0 +1,5 @@
+package failgood.internal.util
+
+class StringUtils
+
+internal fun pluralize(count: Int, item: String) = if (count == 1) "1 $item" else "$count ${item}s"
